+++ R/StoragehubManager.R
@@ -24,8 +24,8 @@ StoragehubManager <-  R6Class("StoragehubManager",
     user_profile = NULL,
     user_workspace = NULL,
     url_icproxy = "https://registry.d4science.org/icproxy/gcube/service/GCoreEndpoint/DataAccess/StorageHub",
-    url_homelibrary = "https://api.d4science.org/social-networking-library-ws/rest/2",
-    url_storagehub = NULL,
+    url_homelibrary = "https://api.d4science.org/rest/2",
+    url_storagehub = "https://api.d4science.org/workspace",
     token_type = NULL,
     supportedTokenTypes = c("gcube", "jwt"),
     #utils
@@ -37,8 +37,7 @@ StoragehubManager <-  R6Class("StoragehubManager",
   ),
   public = list(
     
-    #'@description Method is used to instantiate the \link{StoragehubManager}. By default,
-    #' the url is inherited through D4Science Icproxy service.
+    #'@description Method is used to instantiate the \link{StoragehubManager}.
     #'@param token user access token
     #'@param token_type token type, either 'gcube' (default) or 'jwt'
     #'@param logger logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" (for complete 
@@ -55,9 +54,8 @@ StoragehubManager <-  R6Class("StoragehubManager",
         }
         private$token_type <- token_type
         private$keyring_backend <- keyring:::known_backends[[keyring_backend]]$new()
-        private$keyring_service = paste0("d4storagehub4R@", private$url_icproxy)
+        private$keyring_service = paste0("d4storagehub4R@", private$url_storagehub)
         private$keyring_backend$set_with_value(service = private$keyring_service, username = "d4storagehub4R", password = token)
-        self$fetchWSEndpoint()
         self$fetchUserProfile()
       }else{
         self$INFO("Successfully connected to Storage hub as anonymous user")
@@ -89,6 +87,7 @@ StoragehubManager <-  R6Class("StoragehubManager",
     },
     
     #'@description Fetches the workspace endpoint from the D4Science ICProxy service
+    #'@note Deprecated
     fetchWSEndpoint = function(){
       self$INFO("Fetching workspace endpoint...")
       icproxy_req <- switch(private$token_type,
@@ -129,9 +128,9 @@ StoragehubManager <-  R6Class("StoragehubManager",
       }
     },
     
-    #'@description Get workspace root ID
-    #'@return the workspace root ID, as \code{character}
-    getWSRootID = function(){
+    #'@description Get workspace root
+    #'@return the workspace root, as \code{list}
+    getWSRoot = function(){
       outroot <- NULL
       root_req <- switch(private$token_type,
         "gcube" = {
@@ -145,19 +144,27 @@ StoragehubManager <-  R6Class("StoragehubManager",
       )
       if(!is.null(root_req)){
         rootDoc <- httr::content(root_req)
-        outroot <- rootDoc$item$id
+        outroot <- rootDoc$item
       }
       return(outroot)
     },
     
-    #'@description Get workspace item ID given a \code{folderPath} in a parent folder
+    #'@description Get workspace root ID
+    #'@return the workspace root ID, as \code{character}
+    getWSRootID = function(){
+      outroot <- self$getWSRoot()
+      return(outroot$id)
+    },
+    
+    #'@description Get workspace item given a \code{itemPath} in a parent folder
     #'@param parentFolderID parent folder ID
-    #'@param folderPath folder path
-    #'@return the workspace item ID, \code{NULL} if no workspace item existing
-    getWSItemID = function(parentFolderID = NULL, folderPath){
-      elements <- self$listWSItems(parentFolderID = parentFolderID)
+    #'@param itemPath item path
+    #'@param showHidden show hidden files
+    #'@return the workspace item, \code{NULL} if no workspace item existing
+    getWSItem = function(parentFolderID = NULL, itemPath, showHidden = FALSE){
+      elements <- self$listWSItems(parentFolderID = parentFolderID, showHidden = showHidden)
       
-      wsItemID <- NULL
+      wsItem <- NULL
       if(length(elements)>0) for (i in 1:nrow(elements)){
         el <- elements[i,]
         if (!startsWith(el$path,"/Share/")){
@@ -165,52 +172,80 @@ StoragehubManager <-  R6Class("StoragehubManager",
           if(startsWith(el_path, self$getUserWorkspace())){
             el_path <- unlist(strsplit(el_path, paste0(self$getUserWorkspace(),"/")))[2]
           }
-          if(startsWith(folderPath, self$getUserWorkspace())){
-            folderPath = unlist(strsplit(folderPath, paste0(self$getUserWorkspace(),"/")))[2]
+          if(startsWith(itemPath, self$getUserWorkspace())){
+            itemPath = unlist(strsplit(itemPath, paste0(self$getUserWorkspace(),"/")))[2]
           }
-          if (folderPath == el_path || folderPath == paste0("/",el$path)){
-            wsItemID = el$id
+          if (itemPath == el_path || itemPath == paste0("/",el$path)){
+            wsItem = el
             break
           }
         }else{
-          path.parts = unlist(strsplit(folderPath,"/"))
+          path.parts = unlist(strsplit(itemPath,"/"))
           folder <- path.parts[length(path.parts)]
           el_ws_parts = unlist(strsplit(el$path, "/"))
           el_ws_folder = el_ws_parts[length(el_ws_parts)]
           if (folder == el_ws_folder){
-            wsItemID<-el$id
+            wsItem <- el
             break
           }
         }
       }
-      return(wsItemID)
-      
+      return(wsItem)
     },
     
-    #'@description Get workspace element ID given a \code{folderPath} in a parent folder
+    #'@description Get workspace item ID given a \code{itemPath} in a parent folder
     #'@param parentFolderID parent folder ID
-    #'@param folderPath folder path
-    #'@return the workspace element ID, \code{NULL} if no workspace item existing
-    #'@note Deprecated, use \code{getWSItemID}
-    getWSElementID = function(parentFolderID = NULL, folderPath){
-      self$WARN("Method 'getWSElementID' is deprecated, use method 'getWSItemID' instead!")
-      self$getWSItemID(parentFolderID = parentFolderID, folderPath = folderPath)
+    #'@param itemPath item path
+    #'@param showHidden show hidden files
+    #'@return the workspace item ID, \code{NULL} if no workspace item existing
+    getWSItemID = function(parentFolderID = NULL, itemPath, showHidden = FALSE){
+      item <- self$getWSItem(parentFolderID = parentFolderID, itemPath = itemPath, showHidden = showHidden)
+      return(item$id)
+    },
+    
+    #'@description Get VRE Folder
+    #'@return the VRE folder, as \code{list}
+    getWSVREFolder = function(){
+      outroot <- NULL
+      root_req <- switch(private$token_type,
+       "gcube" = {
+         rootUrl <- paste0(private$url_storagehub, "/vrefolder?exclude=hl:accounting&gcube-token=", self$getToken())
+         httr::GET(rootUrl)
+       },
+       "jwt" = {
+         rootUrl <- paste0(private$url_storagehub, "/vrefolder?exclude=hl:accounting")
+         httr::GET(rootUrl, httr::add_headers("Authorization" = paste("Bearer", self$getToken())))
+       }
+      )
+      if(!is.null(root_req)){
+        rootDoc <- httr::content(root_req)
+        outroot <- rootDoc$item
+      }
+      return(outroot)
+    },
+    
+    #'@description Get VRE Folder ID
+    #'@return the VRE folder ID, as \code{character}
+    getWSVREFolderID = function(){
+      outroot <- self$getWSVREFolder()
+      return(outroot$id)
     },
     
     #'@description Lists workspace items given a parentFolder ID
     #'@param parentFolderID parent folder ID
+    #'@param showHidden show hidden files
     #'@return an object of class \code{data.frame}
-    listWSItems = function(parentFolderID = NULL){
+    listWSItems = function(parentFolderID = NULL, showHidden = FALSE){
       outlist <- NULL
       if(is.null(parentFolderID)) parentFolderID = self$getWSRootID()
       listElementsUrl = paste0(private$url_storagehub, "/items/", parentFolderID, "/children?exclude=hl:accounting")
       list_req <- switch(private$token_type,
         "gcube" = {
-          listElementsUrl <- paste0(listElementsUrl, "&gcube-token=", self$getToken())
+          listElementsUrl <- paste0(listElementsUrl, "&showHidden=", tolower(showHidden), "&gcube-token=", self$getToken())
           httr::GET(listElementsUrl)
         },
         "jwt" = {
-          httr::GET(listElementsUrl, httr::add_headers("Authorization" = paste("Bearer", self$getToken())))
+          httr::GET(paste0(listElementsUrl,"&showHidden=",tolower(showHidden)), httr::add_headers("Authorization" = paste("Bearer", self$getToken())))
         }
       )
       if(!is.null(list_req)){
@@ -220,43 +255,28 @@ StoragehubManager <-  R6Class("StoragehubManager",
       return(outlist)
     },
     
-    #'@description Lists workspace elements given a parentFolder ID
-    #'@param parentFolderID parent folder ID
-    #'@return an object of class \code{data.frame}
-    #'@note Deprecated, use \code{listWSItems}
-    listWSElements = function(parentFolderID = NULL){
-      self$WARN("Method 'listWSElements' is deprecated, use method 'listWSItems' instead!")
-      self$listWSItems(parentFolderID = parentFolderID)
-    },
-    
     #'@description Lists workspace items given a folder path
     #'@param folderPath folder path where to list items
+    #'@param showHidden show hidden files
     #'@return an object of class \code{data.frame}
-    listWSItemsByPath = function(folderPath){
-      folderID <- self$searchWSItemID(itemPath = folderPath)
+    listWSItemsByPath = function(folderPath, showHidden = FALSE){
+      folderID <- self$searchWSItemID(itemPath = folderPath, showHidden = showHidden)
       if(is.null(folderID)) return(NULL)
-      self$listWSItems(parentFolderID = folderID)
+      self$listWSItems(parentFolderID = folderID, showHidden = showHidden)
     },
     
-    #'@description Lists workspace elements given a folder path
-    #'@param folderPath folder path where to list elements
-    #'@return an object of class \code{data.frame}
-    #'@note Deprecated, use \code{listWSItemsByPath}
-    listWSElementsByPath = function(folderPath){
-      self$WARN("Method 'listWSElementsByPath' is deprecated, use method 'listWSItemsByPath' instead!")
-      self$listWSItemsByPath(folderPath = folderPath)
-    },
-    
-    #'@description Searches for a workspace item ID given a item path
+    #'@description Searches for a workspace item given a item path
     #'@param itemPath path of the item
-    #'@return the item ID, \code{NULL} if nothing found
-    searchWSItemID = function(itemPath){
-      rootID = self$getWSRootID()
-      root <- self$getUserWorkspace()
+    #'@param includeVreFolder search also in VRE folder
+    #'@param showHidden show hidden files
+    #'@return the item, \code{NULL} if nothing found
+    searchWSItem = function(itemPath, includeVreFolder = TRUE, showHidden = FALSE){
+      root = self$getWSRoot()
+      rootPath <- self$getUserWorkspace()
       
       if (itemPath==paste("/Home/",self$getUserProfile()$username,"/Workspace",sep="") || 
           itemPath==paste("/Home/",self$getUserProfile()$username,"/Workspace/",sep="")){
-        return(rootID)
+        return(root)
       }
       
       path.splits <- unlist(strsplit(itemPath, "Workspace"))
@@ -264,40 +284,51 @@ StoragehubManager <-  R6Class("StoragehubManager",
       allsubfolders = unlist(strsplit(itemPath, "/"))
       allsubfolders = allsubfolders[nzchar(allsubfolders)]
       
-      parentID = rootID
-      parentPath = root
+      parent = root
+      parentPath = rootPath
       for (subfolder in allsubfolders){
         parentPath = paste0(parentPath,"/",subfolder)
-        parentID = self$getWSItemID(parentFolderID = parentID, folderPath = parentPath)
+        parent = self$getWSItem(parentFolderID = parent$id, itemPath = parentPath, showHidden = showHidden)
       }
-      return(parentID)
+      if(includeVreFolder) if(is.null(parent$id)){
+        vrefolder <- self$getWSVREFolder()
+        parent = vrefolder
+        parentPath = vrefolder$path
+        for (subfolder in allsubfolders){
+          parentPath = paste0(parentPath,"/",subfolder)
+          parent = self$getWSItem(parentFolderID = parent$id, itemPath = parentPath, showHidden = showHidden)
+        }
+      }
+      return(parent)
     },
     
-    #'@description Searches for a workspace element ID given a item path
-    #'@param folderPath path of the element
-    #'@return the item ID, \code{NULL} if nothing found
-    #'@note Deprecated, use \code{searchWSItemID}
-    searchWSFolderID = function(folderPath){
-      self$WARN("Method 'searchWSFolderID' is deprecated, use method 'searchWSItemID' instead!")
-      return(self$searchWSItemID(itemPath = folderPath))
+    #'@description Searches for a workspace item ID given a item path
+    #'@param itemPath path of the item
+    #'@param includeVreFolder search also in VRE folder
+    #'@param showHidden show hidden files
+    #'@return the item, \code{NULL} if nothing found
+    searchWSItemID = function(itemPath, includeVreFolder = TRUE, showHidden = FALSE){
+      item <- self$searchWSItem(itemPath = itemPath, includeVreFolder = includeVreFolder, showHidden = showHidden)
+      return(item$id)
     },
     
     #'@description Creates a folder, given a folder path, a folder name/description. By default \code{recursive = TRUE} meaning 
     #'    that a folder path matching nested folders will trigger all nested folders. Setting \code{recursive = FALSE}, the
     #'    folder creation will work only if the folder path matches an existing folder. The \code{hidden} (default 
-    #'    \code{FALSE}) argument can be used to set hidden folders on the workspace.
+    #'    \code{FALSE}) argument can be used to set hidden folders on the workspace. Using \code{folderID}, \code{recursive} will be
+    #'    set to \code{FALSE}.
     #'@param folderPath parent folder path where to create the folder
+    #'@param folderID parent folder ID where to create the folder
     #'@param name name of the folder
     #'@param description description of the folder
     #'@param hidden hidden, default is \code{FALSE}
     #'@param recursive recursive, default is \code{TRUE}
     #'@return the ID of the created folder
-    createFolder = function(folderPath = NULL, name, description = "", 
+    createFolder = function(folderPath = NULL, folderID = NULL, name, description = "", 
                             hidden = FALSE, recursive = TRUE){
       self$INFO(sprintf("Creating folder '%s at '%s'...", name, folderPath))
-      if(is.null(folderPath)) folderPath = self$getUserWorkspace()
-      
-      
+      if(is.null(folderPath) && is.null(folderID)) folderPath = self$getUserWorkspace()
+      if(!is.null(folderID)) recursive <- FALSE
       if(recursive){
         self$INFO("Recursive mode - Check parent folder(s) and create them if missing...")
         folder_paths <- data.frame(folderPath = character(0), name = character(0), stringsAsFactors = FALSE)
@@ -332,7 +363,7 @@ StoragehubManager <-  R6Class("StoragehubManager",
         for(i in 1:nrow(folder_paths)){
           folder_path <- folder_paths[i,]
           self$INFO(sprintf("Search for an existing folder '%s'", file.path(folder_path$folderPath, folder_path$name)))
-          folderID <- self$searchWSItemID(itemPath = file.path(folder_path$folderPath, folder_path$name))
+          folderID <- self$searchWSItemID(itemPath = file.path(folder_path$folderPath, folder_path$name), showHidden = TRUE)
           if(is.null(folderID)){
             self$INFO(sprintf("Folder '%s' does not exist, we create it...", file.path(folder_path$folderPath, folder_path$name)))
             folderID <- self$createFolder(
@@ -347,7 +378,9 @@ StoragehubManager <-  R6Class("StoragehubManager",
         return(folderID)
         
       }else{
-        pathID = self$searchWSItemID(itemPath = folderPath)
+        pathID = NULL
+        if(!is.null(folderPath)) pathID = self$searchWSItemID(itemPath = folderPath, showHidden = TRUE)
+        if(!is.null(folderID)) pathID = folderID
         if(is.null(pathID)){
           errMsg <- sprintf("No folder for path '%s'", folderPath)
           self$ERROR(errMsg)
@@ -416,17 +449,22 @@ StoragehubManager <-  R6Class("StoragehubManager",
     #'@description  Uploads a file to a folder (given a folder path). The argument \code{description} can be used to further describe the
     #'    file to upload. The argument \code{archive} (default = FALSE) indicates the type of item (FILE or ARCHIVE) to be uploaded.
     #'@param folderPath folder path where to upload the file
+    #'@param folderID folder ID where to upload the file
     #'@param file file to upload
     #'@param description file description, default would be the file basename
     #'@param archive archive, default is \code{FALSE} 
     #'@return the ID of the uploaded file   
-    uploadFile = function(folderPath = NULL, file, description = basename(file), archive = FALSE){
+    uploadFile = function(folderPath = NULL, folderID = NULL, file, description = basename(file), archive = FALSE){
       self$INFO(sprintf("Uploading file '%s' at '%s'...", file, folderPath))
-      if(is.null(folderPath)) folderPath = self$getUserWorkspace()
+      if(is.null(folderPath) && is.null(folderID)) folderPath = self$getUserWorkspace()
       
       name = basename(file)
-      folderPath <- private$normalizeFolderPath(folderPath)
-      pathID <- self$searchWSItemID(itemPath = folderPath)
+      pathID <- NULL
+      if(!is.null(folderPath)){
+        folderPath <- private$normalizeFolderPath(folderPath)
+        pathID <- self$searchWSItemID(itemPath = folderPath, showHidden = TRUE)
+      }
+      if(!is.null(folderID)) pathID <- folderID
       if(is.null(pathID)){
         errMsg <- sprintf("No folder for path '%s'", folderPath)
         self$ERROR(errMsg)
@@ -512,7 +550,7 @@ StoragehubManager <-  R6Class("StoragehubManager",
     #'@return \code{TRUE} if deleted, \code{FALSE} otherwise
     deleteItem = function(itemPath, force = FALSE){
       deleted <- FALSE
-      pathID <- self$searchWSItemID(itemPath = itemPath)
+      pathID <- self$searchWSItemID(itemPath = itemPath, showHidden = TRUE)
       if(!is.null(pathID)){
         delete_url <- sprintf("%s/items/%s", private$url_storagehub, pathID)
         if(force){
@@ -603,6 +641,85 @@ StoragehubManager <-  R6Class("StoragehubManager",
         shared <- FALSE
       }
       return(shared)
+    },
+    
+    #'@description unshare an item
+    #'@param itemPath item path
+    #'@param users users
+    #'@return \code{TRUE} if unshared, \code{FALSE} otherwise
+    unshareItem = function(itemPath, users){
+      unshared <- FALSE
+      the_users <- sapply(users, function(x){list(users = x)})
+      names(the_users) <- rep("users", length(the_users))
+      body <- c(the_users)
+      
+      pathID <- self$searchWSItemID(itemPath = itemPath)
+      if(!is.null(pathID)){
+        unshare_url <- sprintf("%s/items/%s/unshare", private$url_storagehub, pathID)
+        unshared_req <- switch(private$token_type,
+         "gcube" = {
+           unshare_url <-paste0(unshare_url, "?gcube-token=", self$getToken())
+           httr::PUT(
+             unshare_url, 
+             encode = "multipart", 
+             body = body
+           )
+         },
+         "jwt" = {
+           httr::PUT(
+             unshare_url, 
+             httr::add_headers("Authorization" = paste("Bearer", self$getToken())),
+             encode = "multipart",
+             body = body
+           )
+         }
+        )
+        if(!is.null(unshared_req)) if(httr::status_code(unshared_req)==200){
+          unshared <- TRUE
+        }
+      }else{
+        self$WARN(sprintf("No item for path '%s'. Nothing to share!", itemPath))
+        unshared <- FALSE
+      }
+      return(unshared)
+    },
+    
+    #'@description Download item
+    #'@param item item
+    #'@param wd working directory where to download the item
+    downloadItem = function(item = NULL, wd = NULL){
+      if(is.null(wd)) wd <- getwd()
+      link <- NULL
+      pathID <- item$id
+      link_url <- sprintf("%s/items/%s/download?exclude=hl:accounting", private$url_storagehub, pathID)
+      pl_req <- switch(private$token_type,
+       "gcube" = {
+         link_url <-paste0(link_url, "&gcube-token=", self$getToken())
+         httr::GET(link_url)
+       },
+       "jwt" = {
+         httr::GET(link_url, httr::add_headers("Authorization" = paste("Bearer", self$getToken())))
+       }
+      )
+      if(!is.null(pl_req)){
+        data <- httr::content(pl_req, type = "raw")
+        writeBin(data, file.path(wd, basename(item$path)))
+        return(file.path(wd, basename(item$path)))
+      }
+      return(NULL)
+    },
+    
+    #'@description Download item by path
+    #'@param path path
+    #'@param wd working directory where to download the item
+    downloadItemByPath = function(path, wd = NULL){
+      item = self$searchWSItem(itemPath = path, showHidden = TRUE)
+      if(is.null(item)){
+        errMsg <- sprintf("No item for path '%s'", path)
+        self$ERROR(errMsg)
+        stop(errMsg)
+      }
+      self$downloadItem(item = item, wd = wd)
     },
     
     #'@description Get public file link